requirements.txt
numpy
scipy
pandas
python-telegram-bot
python-binance